refactor: 프록시 패턴을 활용하여 비즈니스로직에서 트랜잭션 로직 제거
