+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,46 +1,13 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserHistory;
-import org.springframework.jdbc.support.TransactionTemplate;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionTemplate transactionTemplate;
+    User findById(final long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final TransactionTemplate transactionTemplate) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = transactionTemplate;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var user = findById(id);
-        user.changePassword(newPassword);
-
-        transactionTemplate.execute(() -> {
-            updateUser(user);
-            writeUserHistory(new UserHistory(user, createBy));
-        });
-    }
-
-    private void updateUser(final User user) {
-        userDao.update(user);
-    }
-
-    private void writeUserHistory(final UserHistory userHistory) {
-        userHistoryDao.log(userHistory);
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
