+++ app/src/test/java/com/techcourse/service/TxUserServiceTest.java
@@ -1,7 +1,7 @@
 package com.techcourse.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.dao.UserDao;
@@ -9,19 +9,19 @@ import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.TransactionTemplate;
 
-@Disabled
-class UserServiceTest {
+
+class TxUserServiceTest {
 
     private JdbcTemplate jdbcTemplate;
     private UserDao userDao;
     private UserHistoryDao userHistoryDao;
     private TransactionTemplate transactionTemplate;
+    private UserService appUserService;
 
     @BeforeEach
     void setUp() {
@@ -29,6 +29,7 @@ class UserServiceTest {
         this.userDao = new UserDao(jdbcTemplate);
         this.userHistoryDao = new UserHistoryDao(jdbcTemplate);
         this.transactionTemplate = new TransactionTemplate(DataSourceConfig.getInstance());
+        this.appUserService = new AppUserService(userDao, userHistoryDao);
 
         DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
         final var user = new User("gugu", "password", "hkkang@woowahan.com");
@@ -37,7 +38,7 @@ class UserServiceTest {
 
     @Test
     void testChangePassword() {
-        final var userService = new UserService(userDao, userHistoryDao, transactionTemplate);
+        final var userService = new TxUserService(appUserService, transactionTemplate);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -52,7 +53,8 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionTemplate);
+        final var mockAppUserService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(mockAppUserService, transactionTemplate);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
